issue: 4315534 Remove snd_queuelen
snd_queuelen limits TCP unsent queue by number of pbufs. This is
additional restriction to the TCP send buffer and reduces pbuf objects
exhaustion. On the other hand, snd_queuelen add complexity and isn't
handled around EPOLLOUT and similar events. Therefore, situation is
possible when a socket is ready for write, but it's not possible
due to the queue limit and this repeats wasting 100% of CPU.

For the copy scenario, we can copy data to the tail pbuf, so the queue
doesn't grow unexpectedly and the send buffer becomes sufficient
limiting factor. Zerocopy scenario uses dynamically expanded pool of
small objects.

Mixed scenario of 1 small copy data and 1 small zerocopy data still can
be problematic and grow to a big number of registered buffers which are
limited. This case can be handled by rework of the TCP queue.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>