issue: 4315534 tcp_timers_collection cleanup
Remove free_tta_resources().

Don't expect a destroyed socket in a timer collection list. Timer
insertion and timer ticks are serialized. Besides, we have a sanity
check that a socket doesn't appear in the timers collection twice.
Timer can be removed after a tick in case of local timers, however,
this affects only the previous element of the std::list and cannot
lead to a duplicate socket handling within the loop.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>